#7 upadate: Comment - login in  setting #7 update: Schedule - log in setting

+++ src/main/java/com/schedule/service/CommentService.java
@@ -10,14 +10,16 @@ import com.schedule.repository.ScheduleRepository;
 import com.schedule.repository.UserRepository;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.Optional;
 
-import static com.schedule.entity.Comment.createByRequestDtoAtComment; // TODO 이런식으로 된다!
+import static com.schedule.entity.Comment.createComment; // TODO 이런식으로 된다!
 
+@Slf4j
 @Service
 @AllArgsConstructor
 public class CommentService {
@@ -26,22 +28,31 @@ public class CommentService {
     private final UserRepository userRepository;
     private final CommentRepository commentRepository;
 
-    public CommentResponseDto createCommet(CommentRequestDto dto, Long scheduleId) {
+    @Transactional
+    public CommentResponseDto createCommet(CommentRequestDto dto, String email , Long scheduleId) {
 
-        if(userRepository.findByUserName(dto.getUserName()) == null){
+        Optional<User> byEmail = userRepository.findByEmail(email);
+
+        log.info("test");
+
+        User user = byEmail.get();
+
+        if(!user.getUserName().equals(dto.getUserName())) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND,"본인 계정에 이름을 입력해주세요.");
         }
 
-        Result result = findByDtoAndByScheduleId(dto, scheduleId);
+        Optional<Schedule> findByIdSchedule = scheduleRepository.findById(scheduleId);
+        findByIdSchedule.orElseThrow(()-> new ResponseStatusException(HttpStatus.NOT_FOUND));
+        Schedule schedule = findByIdSchedule.get();
 
-        Comment byRequestDtoAtComment =  createByRequestDtoAtComment(dto.getComment(), result.user(), result.schedule()); // TODO 신기!
+        Comment createComment =  createComment(dto.getComment(),user,schedule); // TODO 신기!
 
-        commentRepository.save(byRequestDtoAtComment);
+        commentRepository.save(createComment);
 
-        return new CommentResponseDto(byRequestDtoAtComment.getComment());
+        return new CommentResponseDto(createComment.getComment());
     }
 
-
+    @Transactional
     public CommentResponseDto findByIdComment(Long commentId) {
 
         Optional<Comment> optionalFindById = commentRepository.findById(commentId);
@@ -54,52 +65,41 @@ public class CommentService {
     }
 
     @Transactional
-    public CommentResponseDto modifyByIdComment(CommentRequestDto dto, Long commentId) {
+    public CommentResponseDto modifyByIdComment(CommentRequestDto dto, Long commentId ,String email ) {
 
-        Optional<Comment> optionalFindById = commentRepository.findById(commentId);
+        log.info("본인 계정임을 확인합니다.");
 
-        optionalFindById.orElseThrow(()->new ResponseStatusException(HttpStatus.NOT_FOUND));
+        Comment comment = selfCommentToCheck(commentId);
 
-        Comment comment = optionalFindById.get();
+        if(!comment.getUser().getEmail().equals(email)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,"본인 계정에 작성물만 수정할 수 있습니다.");
+        }
 
         comment.setComment(dto.getComment());
 
         return new CommentResponseDto(comment.getComment());
     }
 
+    @Transactional
+    public void DeleteByIdComment(Long commentId,String email) {
 
-    public void DeleteByIdComment(Long commentId) {
+        log.info("본인 계정임을 확인합니다.");
+
+        Comment comment = selfCommentToCheck(commentId);
+
+        if(!comment.getUser().getEmail().equals(email)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND,"본인 계정에 작성물만 삭제할 수 있습니다.");
+        }
 
         commentRepository.deleteById(commentId);
-
     }
 
+    private Comment selfCommentToCheck(Long commentId) {
+        Optional<Comment> optionalFindById = commentRepository.findById(commentId);
 
-    private record Result(User user, Schedule schedule) {
+        optionalFindById.orElseThrow(()->new ResponseStatusException(HttpStatus.NOT_FOUND));
+
+        Comment comment = optionalFindById.get();
+        return comment;
     }
-
-    /**
-     * scheduleId 로 ScheduleRepository 의 findById 사용, schedule 객체 찾기
-     * CommentRequestDto 에 dto.getUserName 로, UserRepository 의 findByUserName 사용, user 객체 찾기
-     *
-     * @param dto
-     * @param scheduleId
-     * @return 데이터베이스에 저장된 user 객체와 schedule 객체를 반환
-     */
-    private Result findByDtoAndByScheduleId(CommentRequestDto dto, Long scheduleId) {
-
-        // 해당 과정은 사실 로그인 인증하고 들어온 것이라 확인할 필요가 없음
-        Optional<User> findByUserName = userRepository.findByUserName(dto.getUserName());
-        findByUserName.orElseThrow(()-> new ResponseStatusException(HttpStatus.NOT_FOUND));
-        User user = findByUserName.get();
-
-        // 해당 과정도 controller 에서 pathvaild 를 통해 오기 때문에 의미가 없을 듯?
-        Optional<Schedule> findByIdSchedule = scheduleRepository.findById(scheduleId);
-        findByIdSchedule.orElseThrow(()-> new ResponseStatusException(HttpStatus.NOT_FOUND));
-        Schedule schedule = findByIdSchedule.get();
-
-        Result result = new Result(user, schedule);
-        return result;
-    }
-
 }
